pyproject.toml
[project]
name = "error-demo"
version = "0.1.0"
description = "Demonstrator for tab-err, conformal data cleaning, MechDetect, and XAI."
readme = "README.md"
requires-python = ">=3.10"

authors = [
  { name = "Nicholas Chandler", email = "nich1834@bht-berlin.de" },
]

dependencies = [
  "tab-err",
  "streamlit",
  "itables",
  "plotly",
  "openpyxl",
  "numpy>=2.0.2",
  "openml>=0.15.1",
  "pandas>=2.3.3",
  "quantile-forest>=1.4.1",
  "scikit-learn>=1.6.1",
  "typer-slim>=0.20.0",
  "conformal-data-cleaning @ git+https://github.com/calgo-lab/conformal-data-cleaning.git@refactoring"
]